apply patch from tvainika
git-svn-id: file:///home/lennart/svn/public/avahi/trunk@331 941a03a8-eaeb-0310-b9a0-b1bbd8fe43fe

+++ avahi-core/dns.h
@@ -47,7 +47,7 @@ void avahi_dns_packet_set_field(AvahiDnsPacket *p, unsigned idx, uint16_t v);
 uint16_t avahi_dns_packet_get_field(AvahiDnsPacket *p, unsigned idx);
 void avahi_dns_packet_inc_field(AvahiDnsPacket *p, unsigned idx);
 
-uint8_t *avahi_dns_packet_extend(AvahiDnsPacket *p, unsigned l);
+uint8_t *avahi_dns_packet_extend(AvahiDnsPacket *p, size_t l);
 
 uint8_t *avahi_dns_packet_append_uint16(AvahiDnsPacket *p, uint16_t v);
 uint8_t *avahi_dns_packet_append_uint32(AvahiDnsPacket *p, uint32_t v);
